Update src/main/java/co/fineants/api/domain/member/controller/SignUpRestController.java
I don't think I need to put a if statement. Because I add code in DefaultAmazonS3Service. code is null check if statement.
deleteProfileImageFile don't have issuees if it's method fail.

+++ src/main/java/co/fineants/api/domain/member/controller/SignUpRestController.java
@@ -70,7 +70,9 @@ public class SignUpRestController {
 			signupService.signup(member);
 		} catch (BusinessException exception) {
 			log.warn("BusinessException occurred during signup: {}", exception.getMessage(), exception);
-			signupService.deleteProfileImageFile(profileUrl);
+			if (profileUrl != null) {
+				signupService.deleteProfileImageFile(profileUrl);
+			}
 			throw new SignupException(exception);
 		}
